Fixed the particles shader, GUIEditBox and GUIScrollBar render

+++ src/client/core/client.cpp
@@ -908,7 +908,7 @@ struct TextureUpdateArgs {
 void Client::showUpdateProgressTexture(void *args, u32 progress, u32 max_progress)
 {
 		TextureUpdateArgs* targs = (TextureUpdateArgs*) args;
-		u16 cur_percent = std::ceil(progress / max_progress * 100.f);
+    u16 cur_percent = std::ceil(progress / (f32)max_progress * 100.f);
 
 		// update the loading menu -- if necessary
 		bool do_draw = false;
